changes
Old interviwew game is still up but all the changes are in a new scene called InterviewGame. Feel free to work on it, I'll work on it during the week as well. All the current scripts for the new scene are in a scprit folder named interview game (or somehthing similar)
Hit me up if you have questions I tried to make things obvious with the names and comments. All the buttons should be set up minus the characters.